requirements.txt
librosa
torchvision
mir_eval
pypesq
pystoi
h5py
pytorch
torchvision
torchaudio
mmengine
mmcv
av
facenet_pytorch
opencv-python
